backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn==1.7.2
joblib
openpyxl
xlrd
requests
langchain-groq
python-dotenv
imbalanced-learn
